fix(clear-macro): split tombstone expiry by cause, and finish the pending/terminal surfaces
The plan is self-contradictory about tombstone expiry, and this takes the narrower reading.
§10.2 says a guard releases only on a positive answer and "never on a timeout, a deadline, a
5xx, or an unreachable provider", then says a tombstone "expires on its own at
`validBefore + grace`". Those cannot both hold when the provider was simply unreachable:
passing `validBefore` proves the payload cannot land in future, not that it never landed
during the outage, so releasing there IS releasing on a timeout.

Tombstones now record why they exist. A confirmed 404 expires — the provider answered, and
past the window the payload cannot land. An unreachable provider does not, and waits for a
positive answer or the user's explicit override, which is exactly what that override is for.
Flagged for Kaspar rather than decided quietly.

Also finishes two surfaces the plan asks for that were only half-built:

- §11 wanted Review in Safe, Cancel and a keepable execution id in the DIALOG, not only in a
  notification. Extracted into `SafeRelayPendingActions` and used by both, so they cannot
  drift. The id matters here because it is the only handle on a live intent — there is no
  lookup by signer — so a user who clears storage can otherwise neither cancel it nor safely
  write that action directly until it expires.
- §7.4's terminal copy only reached the background recovery path. The in-session poll (a
  1-of-1 Safe, or any EOA relay) still rendered the raw provider string, so a cancellation, a
  preflight revert and a real failure all read the same — and the old wording offered "safe to
  try again", which is not always true.
