+++ src/features/clearMacro/SafeRelayPendingToast.tsx
@@ -1,8 +1,6 @@
-import { FC, useState } from "react";
-import { Button, Link, Stack, Tooltip, Typography } from "@mui/material";
-import ContentCopyRoundedIcon from "@mui/icons-material/ContentCopyRounded";
-import OpenInNewRoundedIcon from "@mui/icons-material/OpenInNewRounded";
-import { useCancelRelayExecution } from "./useCancelRelayExecution";
+import { FC } from "react";
+import { Stack, Typography } from "@mui/material";
+import { SafeRelayPendingActions } from "./SafeRelayPendingActions";
 
 /**
  * The persistent surface for a Safe gasless request that is waiting on its owners.
@@ -11,10 +9,6 @@ import { useCancelRelayExecution } from "./useCancelRelayExecution";
  * copy is wrong on both halves for a multi-day co-signer wait: nothing is being confirmed yet,
  * and the reason not to retry is not vagueness but a concrete double-spend the user can resolve
  * by cancelling.
- *
- * Everything the user needs to act without this tab survives is here: the expiry, the execution
- * id (which is the ONLY handle on a live intent — there is no lookup by signer), a link to the
- * message in Safe, and Cancel.
  */
 export interface SafeRelayPendingToastProps {
   executionId: string;
@@ -28,12 +22,6 @@ export interface SafeRelayPendingToastProps {
   threshold?: number;
 }
 
-const formatExpiry = (validBefore: number) =>
-  new Date(validBefore * 1000).toLocaleString(undefined, {
-    dateStyle: "medium",
-    timeStyle: "short",
-  });
-
 export const SafeRelayPendingToast: FC<SafeRelayPendingToastProps> = ({
   executionId,
   clientRequestId,
@@ -42,14 +30,9 @@ export const SafeRelayPendingToast: FC<SafeRelayPendingToastProps> = ({
   confirmations,
   threshold,
 }) => {
-  const cancel = useCancelRelayExecution();
-  const [isConfirmingCancel, setIsConfirmingCancel] = useState(false);
-  const [copied, setCopied] = useState(false);
-  const [copyFailed, setCopyFailed] = useState(false);
-
   // Confident copy only when the Transaction Service actually told us a proposal exists.
-  // Otherwise the honest version: we cannot distinguish "signed and closed the modal" from
-  // "declined", and the user can.
+  // Otherwise the honest version: a rejection cannot distinguish "signed and closed the modal"
+  // from "declined" — but the user can, so the decision is theirs.
   const headline =
     confirmations != null
       ? threshold != null
@@ -62,119 +45,21 @@ export const SafeRelayPendingToast: FC<SafeRelayPendingToastProps> = ({
       <Typography variant="body2" translate="yes" sx={{ fontWeight: 500 }}>
         {headline}
       </Typography>
-      <Typography variant="caption" translate="yes" sx={{ color: "text.secondary" }}>
+      <Typography
+        variant="caption"
+        translate="yes"
+        sx={{ color: "text.secondary" }}
+      >
         {confirmations != null
           ? "It will go through on its own once they confirm in Safe — you can close this tab."
           : "If you approved the request, the other owners can confirm it in Safe and this will go through on its own — you can close this tab. If you declined it, cancel below to release it."}
       </Typography>
-      <Typography variant="caption" translate="yes" sx={{ color: "text.secondary" }}>
-        Open until {formatExpiry(validBefore)}.
-      </Typography>
-
-      <Stack
-        direction="row"
-        sx={{ alignItems: "center", gap: 0.5, flexWrap: "wrap" }}
-      >
-        <Typography variant="caption" sx={{ color: "text.secondary" }} translate="no">
-          {executionId}
-        </Typography>
-        <Tooltip
-          title={
-            copyFailed
-              ? "Couldn't copy — select the ID and copy it manually"
-              : copied
-                ? "Copied"
-                : "Copy execution ID"
-          }
-        >
-          <Button
-            size="small"
-            data-cy="safe-relay-copy-execution-id"
-            aria-label="Copy execution ID"
-            sx={{ minWidth: 0, p: 0.5 }}
-            onClick={() => {
-              // Clipboard access can be denied outright; a rejected promise here must not
-              // become an unhandled rejection just because a copy button was pressed.
-              navigator.clipboard
-                ?.writeText(executionId)
-                .then(() => {
-                  setCopied(true);
-                  setTimeout(() => setCopied(false), 2000);
-                })
-                .catch(() => setCopyFailed(true));
-            }}
-          >
-            <ContentCopyRoundedIcon fontSize="inherit" />
-          </Button>
-        </Tooltip>
-      </Stack>
-
-      <Stack direction="row" sx={{ gap: 1, flexWrap: "wrap" }}>
-        {messageLink && (
-          <Link
-            href={messageLink}
-            target="_blank"
-            rel="noopener noreferrer"
-            variant="caption"
-            data-cy="safe-relay-review-in-safe"
-            sx={{ display: "inline-flex", alignItems: "center", gap: 0.25 }}
-          >
-            Review in Safe
-            <OpenInNewRoundedIcon fontSize="inherit" />
-          </Link>
-        )}
-        {!isConfirmingCancel ? (
-          <Button
-            size="small"
-            color="error"
-            data-cy="safe-relay-cancel"
-            onClick={() => setIsConfirmingCancel(true)}
-          >
-            Cancel
-          </Button>
-        ) : (
-          <Stack sx={{ gap: 0.5 }}>
-            <Typography variant="caption" translate="yes">
-              Cancel this gasless request? This can&apos;t be undone.
-            </Typography>
-            <Stack direction="row" sx={{ gap: 1 }}>
-              <Button
-                size="small"
-                color="error"
-                variant="contained"
-                data-cy="safe-relay-cancel-confirm"
-                disabled={cancel.isPending}
-                onClick={() =>
-                  cancel.mutate({ executionId, clientRequestId, messageLink })
-                }
-              >
-                {cancel.isPending ? "Cancelling…" : "Yes, cancel"}
-              </Button>
-              <Button
-                size="small"
-                data-cy="safe-relay-cancel-dismiss"
-                disabled={cancel.isPending}
-                onClick={() => setIsConfirmingCancel(false)}
-              >
-                Keep it
-              </Button>
-            </Stack>
-          </Stack>
-        )}
-      </Stack>
-
-      {cancel.isError && (
-        <Typography
-          variant="caption"
-          color="error"
-          translate="yes"
-          data-cy="safe-relay-cancel-error"
-        >
-          {cancel.error instanceof Error
-            ? cancel.error.message
-            : "The cancellation could not be confirmed."}
-        </Typography>
-      )}
+      <SafeRelayPendingActions
+        executionId={executionId}
+        clientRequestId={clientRequestId}
+        validBefore={validBefore}
+        messageLink={messageLink}
+      />
     </Stack>
   );
 };
